Move schema file and update references
